fix: combination frontend + backend

+++ frontend/src/utils/Api.js
@@ -1,9 +1,13 @@
 export default class Api {
-  constructor({ baseUrl, headers }) {
-    this._baseUrl = baseUrl;
+  constructor({
+    url,
+    headers
+  }) {
+    this._url = url;
     this._headers = headers;
   }
 
+  // Метод проверки ответа сервера
   _getResponseData(res) {
     if (res.ok) {
       return res.json();
@@ -11,69 +15,75 @@ export default class Api {
     return Promise.reject(`Ошибка: ${res.status}`);
   }
 
+  // Получем информацию о пользователе с сервера
   getUserData() {
-    return fetch(`${this._baseUrl}/users/me`,
-    { headers: this._headers }).then((res) => this._getResponseData(res));
+    return fetch(`${this._url}/users/me`, {
+      headers: this._headers,
+    }).then((res) => this._getResponseData(res));
   }
 
-  getInitialCards() {
-    return fetch(`${this._baseUrl}/cards`,
-    { headers: this._headers }).then((res) => this._getResponseData(res));
-  }
-
-  saveUserChanges(data){
-    return fetch(`${this._baseUrl}/users/me`, {
+  // Сохраняем отредактированные данные пользователя на сервере
+  saveUserChanges(data) {
+    return fetch(`${this._url}/users/me`, {
       method: 'PATCH',
+      headers: this._headers,
       body: JSON.stringify({
         name: data.name,
         about: data.about,
       }),
-      headers: this._headers }).then((res) => this._getResponseData(res));
+    }).then((res) => this._getResponseData(res));
   }
 
+  // Обновляем аватар пользователя
+  changedAvatar(data) {
+    return fetch(`${this._url}/users/me/avatar`, {
+      method: 'PATCH',
+      headers: this._headers,
+      body: JSON.stringify({
+        avatar: data.avatar,
+      }),
+    }).then((res) => this._getResponseData(res));
+  }
+
+  // Получаем карточеки с сервера
+  getInitialCards() {
+    return fetch(`${this._url}/cards`, {
+      headers: this._headers,
+    }).then((res) => this._getResponseData(res));
+  }
+
+  // Добавляем новую карточку на сервер
   postNewCard(card) {
-    return fetch(`${this._baseUrl}/cards`, {
+    return fetch(`${this._url}/cards`, {
       method: 'POST',
+      headers: this._headers,
       body: JSON.stringify({
         name: card.name,
         link: card.link,
       }),
-      headers: this._headers }).then((res) => this._getResponseData(res));
+    }).then((res) => this._getResponseData(res));
   }
 
+  // Удаляем карточки пользователя с сервера
   deleteCard(cardId) {
     return fetch(`${this._url}/cards/${cardId}`, {
       method: 'DELETE',
-      headers: this._headers }).then((res) => this._getResponseData(res));
+      headers: this._headers,
+    }).then((res) => this._getResponseData(res));
   }
 
-  changedAvatar(data) {
-    return fetch(`${this._url}/users/me/avatar`, {
-      method: 'PATCH',
-      body: JSON.stringify({
-        avatar: data.avatar,
-      }),
-      headers: this._headers }).then((res) => this._checkResponse(res));
-  }
-
-  likedCard(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+  // Ставим/удаляем лайк карточке
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
         method: 'PUT',
         headers: this._headers,
       }).then((res) => this._getResponseData(res));
-  }
-
-  dislikedCard(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers }).then((res) => this._getResponseData(res));
+    } else {
+      return fetch(`${this._url}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers,
+      }).then((res) => this._getResponseData(res));
+    }
   }
 }
-
-// const api = new Api({
-//   url: 'https://api.mesto-gallery.student.nomoredomainsicu.ru',
-//   headers: {
-//     'Content-Type': 'application/json',
-//     authorization: `Bearer ${localStorage.getItem('jwt')}`,
-//   },
-// })
